Decouple mass-nav groups from formation layout
