+++ mods/showcases/navigation/MassNavWebParityMod/UI/MassNavPlaygroundPanelController.cs
@@ -541,9 +541,9 @@ internal sealed class MassNavWebParityPanelController
             SelectedTeamId: simulation.SelectedTeamId,
             SelectedCount: simulation.SelectedCount,
             SelectionRevision: simulation.SelectionRevision,
-            FormationCount: simulation.FormationRuntime.ActiveGroupCount,
+            FormationCount: simulation.NavGroupRuntime.ActiveGroupCount,
             FormationLabel: simulation.FormationMode.ToString(),
-            FormationRotationDeg: simulation.FormationRuntime.SelectedRotationRadians * (180f / MathF.PI),
+            FormationRotationDeg: simulation.NavGroupRuntime.SelectedRotationRadians * (180f / MathF.PI),
             FlowEnabled: simulation.FlowTuning.Enabled,
             FlowIterations: simulation.FlowTuning.IterationsPerStep,
             FlowStepInterval: simulation.FlowTuning.StepIntervalTicks,
